pyproject.toml
[project]
name = "rag-ai-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Patricia Cobelli",email = "patycobelli@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "chroma (>=0.2.0,<0.3.0)",
    "chroma-datasets (>=0.1.5,<0.2.0)",
    "datasets (>=3.3.2,<4.0.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "langgraph (>=0.3.6,<0.4.0)",
    "langchain-core (>=0.3.44,<0.4.0)",
    "langchain-ollama (>=0.2.3,<0.3.0)",
    "ipython (>=9.0.2,<10.0.0)",
    "langchain-chroma (>=0.2.2,<0.3.0)",
    "langchain-openai (>=0.3.8,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
